+++ libc/include/llvm-libc-macros/link-macros.h
@@ -24,6 +24,7 @@
 #include "../llvm-libc-types/Elf32_Sword.h"
 #include "../llvm-libc-types/Elf32_Sym.h"
 #include "../llvm-libc-types/Elf32_Word.h"
+#include "../llvm-libc-types/Elf32_Xword.h"
 #include "../llvm-libc-types/Elf64_Addr.h"
 #include "../llvm-libc-types/Elf64_Chdr.h"
 #include "../llvm-libc-types/Elf64_Dyn.h"